reduce many codes
