+++ Cargo.toml
@@ -14,14 +14,13 @@
 
 [package]
 name = "executor"
-version = "0.2.0-alpha.1"
+version = "0.2.0-alpha.2"
 authors = ["Rivtower Technologies <contact@rivtower.com>"]
 license = "Apache-2.0"
 edition = "2018"
 
 [dependencies]
 dotenv = "0.13.0"
-log = "0.4"
 cita-logger = "0.1.1"
 tonic = "0.4"
 git-version = "0.3"
@@ -43,13 +42,9 @@ byteorder = { version = "1", default-features = false }
 bincode = "0.8.0"
 rustc-serialize = "0.3"
 lru-cache = "0.1.1"
-bit-set = "0.4"
-rust-crypto = "0.2.34"
-num = "0.1"
 rand = "0.3"
 transient-hashmap = "0.4.0"
 ethabi = "4.2.0"
-enum_primitive = "0.1.1"
 largest-remainder-method = "0.1"
 parking_lot = "0.8"
 ethbloom = "0.6"
@@ -57,20 +52,16 @@ sha3 = { version="0.8", optional=true }
 tiny-keccak = { version="1.4", optional=true }
 hashbrown = { version = "0.3", features = ["rayon"] }
 hasher = { version="0.1" }
+ethereum-types = "0.4"
 
 cita-merklehash = { git = "https://github.com/citahub/cita-common.git", branch = "develop" }
 cita-secp256k1 = { git = "https://github.com/citahub/cita-common.git", branch = "develop" }
-cita-crypto-trait = { git = "https://github.com/citahub/cita-common.git", branch = "develop" }
 rlp = { git = "https://github.com/citahub/cita-common.git", branch = "develop" }
 util = { git = "https://github.com/citahub/cita-common.git", branch = "develop" }
-cita-types = { git = "https://github.com/citahub/cita-common.git", branch = "develop" }
 hashable = { git = "https://github.com/citahub/cita-common.git", branch = "develop" }
 jsonrpc-types = { git = "https://github.com/citahub/cita-common.git", branch = "develop" }
 cita-crypto = { git = "https://github.com/citahub/cita-common.git", branch = "develop" }
 libproto = { git = "https://github.com/citahub/cita-common.git", branch = "develop" }
-snappy = { git = "https://github.com/citahub/cita-common.git", branch = "develop" }
-proof = { git = "https://github.com/citahub/cita-common.git", branch = "develop" }
-pubsub = { git = "https://github.com/citahub/cita-common.git", branch = "develop" }
 
 cita_cloud_proto = { git = "https://github.com/cita-cloud/cita_cloud_proto" }
 
@@ -86,10 +77,9 @@ tempdir = "0.3.7"
 
 [features]
 default = ["sm2", "sm3hash", "hashlib-keccak"]
-secp256k1 = ["cita-crypto/secp256k1", "libproto/secp256k1", "proof/secp256k1"]
-sm2 = ["cita-crypto/sm2", "libproto/sm2", "proof/sm2"]
-sha3hash = ["hashable/sha3hash", "libproto/sha3hash", "proof/sha3hash"]
-blake2bhash = ["hashable/blake2bhash", "libproto/blake2bhash", "proof/blake2bhash"]
-sm3hash = ["hashable/sm3hash", "libproto/sm3hash", "proof/sm3hash"]
+secp256k1 = ["cita-crypto/secp256k1", "libproto/secp256k1"]
+sm2 = ["cita-crypto/sm2", "libproto/sm2"]
+sha3hash = ["hashable/sha3hash", "libproto/sha3hash"]
+sm3hash = ["hashable/sm3hash", "libproto/sm3hash"]
 hashlib-keccak = ["tiny-keccak"]
 hashlib-sha3 = ["sha3"]
