+++ src/core_executor.rs
@@ -18,7 +18,6 @@ pub mod cita_executive;
 pub mod cita_vm_helper;
 pub mod data_provider;
 pub mod libexecutor;
-pub mod storage;
 pub mod tx_gas_schedule;
 
 mod exception;
